+++ lib/faraday_hal_middleware.rb
@@ -1,10 +1,4 @@
 # frozen_string_literal: true
 
-require 'faraday_hal_middleware/version'
-require_relative 'faraday_hal_middleware/request/hal_json'
-require_relative 'faraday_hal_middleware/response/hal_json'
-
-module FaradayHalMiddleware
-  Faraday::Request.register_middleware(hal_json: ::FaradayHalMiddleware::Request::HalJson)
-  Faraday::Response.register_middleware(hal_json: ::FaradayHalMiddleware::Response::HalJson)
-end
+require 'faraday'
+require_relative 'faraday/hal_json'